# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Documentation",
        "description": "This is a sample documentation",
        "termsOfService": "http://laravelapi.test/api/terms",
        "contact": {
            "email": "cmg.web@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://laravelapi.test/api",
            "description": "Laravel API Test Server"
        },
        {
            "url": "http://laravelapi.stage/api",
            "description": "Laravel API Stage Server"
        }
    ],
    "paths": {
        "/products": {
            "get": {
                "tags": [
                    "product"
                ],
                "summary": "List all products",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "How many items to return at one time",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A paged array of products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "product"
                ],
                "summary": "Create a product",
                "operationId": "store",
                "requestBody": {
                    "description": "Store a product",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            }
        },
        "/products/{productId}": {
            "get": {
                "tags": [
                    "product"
                ],
                "summary": "Info for a specific product",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "The id column of the product to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product detail response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "product"
                ],
                "summary": "Update product",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "The id column of the product to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update product",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product updated response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "product"
                ],
                "summary": "Delete a  product",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "The id column of the product to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Product updated response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Product": {
                "title": "Product",
                "description": "Product model",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "id column",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    }
                },
                "type": "object"
            },
            "ApiResponse": {
                "title": "ApiResponse",
                "description": "ApiResponse model",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "type": "object"
                    },
                    "message": {
                        "type": "string"
                    },
                    "errors": {
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "api_token": {
                "type": "apiKey",
                "name": "api_token",
                "in": "query"
            },
            "berarer_token": {
                "type": "http",
                "bearerFormat": "JWT9",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "product",
            "description": "Product tag descripton",
            "externalDocs": {
                "description": "Find out more",
                "url": "http://laravelapi.test/api/documentation/product"
            }
        }
    ],
    "externalDocs": {
        "description": "api_key: x65ifEV4g3ofFpUYxnoChRmteU9sX773yFYyXXMdMKESVXziHiJmT1Zh4DYV",
        "url": "http://laravelapi.test/ext-documentation"
    },
    "security": [
        []
    ]
}